# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Roomly household management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /households/{householdId}/rooms:
    get:
      operationId: listRooms
      tags: [rooms]
      summary: List rooms in household
      parameters:
        - name: householdId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Rooms
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Room"
    post:
      operationId: createRoom
      tags: [rooms]
      summary: Create a room
      parameters:
        - name: householdId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoomInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
  /rooms/{roomId}:
    patch:
      operationId: updateRoom
      tags: [rooms]
      summary: Update a room
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoomUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
    delete:
      operationId: deleteRoom
      tags: [rooms]
      summary: Delete a room
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Expenses ──────────────────────────────────────────────────────────────
...
  /households/{householdId}/rules:
    get:
      operationId: getHouseRules
      tags: [rules]
      summary: Get the house rules document
      parameters:
        - name: householdId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: House rules
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HouseRules"
    put:
      operationId: updateHouseRules
      tags: [rules]
      summary: Update house rules document
      parameters:
        - name: householdId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HouseRulesInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HouseRules"
  /households/{householdId}/rules/sign:
    post:
      operationId: signHouseRules
      tags: [rules]
      summary: Digitally sign the house rules
      parameters:
        - name: householdId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RuleSignatureInput"
      responses:
        "201":
          description: Signed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RuleSignature"

  # ── Payments ──────────────────────────────────────────────────────────────
  /households/{householdId}/payments:
    get:
      operationId: listPayments
      tags: [payments]
      summary: List rent payments
      parameters:
        - name: householdId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Payments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payment"
    post:
      operationId: recordPayment
      tags: [payments]
      summary: Record a rent payment
      parameters:
        - name: householdId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentInput"
      responses:
        "201":
          description: Recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /payments/{paymentId}:
    patch:
      operationId: updatePayment
      tags: [payments]
      summary: Update a payment record
      parameters:
        - name: paymentId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"

  # ── Dashboard (wow endpoints) ─────────────────────────────────────────────
...
components:
  schemas:
...
    Room:
      type: object
      required: [id, householdId, name, monthlyRent, createdAt]
      properties:
        id:
          type: integer
        householdId:
          type: integer
        name:
          type: string
        monthlyRent:
          type: number
        description:
          type: ["string", "null"]
        createdAt:
          type: string
    RoomInput:
      type: object
      required: [name, monthlyRent]
      properties:
        name:
          type: string
          minLength: 1
        monthlyRent:
          type: number
          minimum: 0
        description:
          type: string
    RoomUpdate:
      type: object
      properties:
        name:
          type: string
        monthlyRent:
          type: number
          minimum: 0
        description:
          type: string

    # ── Expense ───────────────────────────────────────────────────────────
...
    HouseRules:
      type: object
      required: [id, householdId, content, version, createdAt, updatedAt]
      properties:
        id:
          type: integer
        householdId:
          type: integer
        content:
          type: string
        version:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
        signatures:
          type: array
          items:
            $ref: "#/components/schemas/RuleSignature"
    HouseRulesInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
          minLength: 1
    RuleSignature:
      type: object
      required: [id, houseRulesId, memberId, memberName, signedAt]
      properties:
        id:
          type: integer
        houseRulesId:
          type: integer
        memberId:
          type: integer
        memberName:
          type: string
        signedAt:
          type: string
    RuleSignatureInput:
      type: object
      required: [memberId]
      properties:
        memberId:
          type: integer

    # ── Payment ───────────────────────────────────────────────────────────
    Payment:
      type: object
      required: [id, householdId, memberId, amount, month, status, createdAt]
      properties:
        id:
          type: integer
        householdId:
          type: integer
        memberId:
          type: integer
        memberName:
          type: string
        amount:
          type: number
        month:
          type: string
        dueDate:
          type: ["string", "null"]
        paidDate:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, paid, late, partial]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    PaymentInput:
      type: object
      required: [memberId, amount, month]
      properties:
        memberId:
          type: integer
        amount:
          type: number
          minimum: 0
        month:
          type: string
        dueDate:
          type: string
        paidDate:
          type: string
        status:
          type: string
          enum: [pending, paid, late, partial]
        notes:
          type: string
    PaymentUpdate:
      type: object
      properties:
        amount:
          type: number
          minimum: 0
        paidDate:
          type: string
        status:
          type: string
          enum: [pending, paid, late, partial]
        notes:
          type: string

    # ── Dashboard ─────────────────────────────────────────────────────────
...