# lib/api-zod/src/generated/types/room.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Roomly household management API
 * OpenAPI spec version: 0.1.0
 */

export interface Room {
  id: number;
  householdId: number;
  name: string;
  monthlyRent: number;
  /** @nullable */
  description?: string | null;
  createdAt: string;
}