# lib/api-zod/src/generated/types/ruleSignatureInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Roomly household management API
 * OpenAPI spec version: 0.1.0
 */

export interface RuleSignatureInput {
  memberId: number;
}